refactor(sync): move unitest for sync to sync_test.rs
