+++ crates/papyrus_sync/src/sources/mod.rs
@@ -5,9 +5,13 @@ mod central_sync_test;
 #[cfg(test)]
 mod central_test;
 
+#[cfg(test)]
+pub(crate) use base_layer::MockBaseLayerSourceTrait;
 pub use base_layer::{
     BaseLayerError, BaseLayerSourceErrorTrait, BaseLayerSourceTrait, EthereumBaseLayerSource,
 };
+#[cfg(test)]
+pub(crate) use central::MockCentralSourceTrait;
 pub use central::{
     CentralError, CentralResult, CentralSource, CentralSourceConfig, CentralSourceTrait,
 };
